+++ src/modules/users/useCases/verifyUser.ts
@@ -30,7 +30,10 @@ export class VerifyUserImpl implements VerifyUser {
       await this.persistentTokenRepo.getByUserIdAndToken(userId, token);
     if (!verificationToken) throw AppError.unauthorizedError();
 
-    await this.userRepo.update(userId, { verified: true });
+    const user = await this.userRepo.getUserById(userId);
+    if (!user) throw AppError.notFoundError("User not found");
+
+    await this.userRepo.update(user, { verified: true });
     await this.persistentTokenRepo.deleteOne(verificationToken);
   }
 }
